extract createBuildClientConfig to client.ts
